Address PR review comments and move tests to pytest
- Warp import now raises RuntimeError instead of silently falling back to PyTorch
- Skip kernel 2 and grad storage when inputs don't require gradients
- Torch profiler uses proper with block + try/finally for safe trace export
- Expose placement_shrink_dist as CLI arg
- Move correctness tests to tests/test_sphere_overlap.py (pytest-compatible)
- Rename script to benchmark_sphere_overlap.py (benchmark-only)
- Update docs to match current function names and removed sync
